+++ application/modules/journal/controllers/PaperController.php
@@ -591,6 +591,11 @@ class PaperController extends PaperDefaultController
             if ($affi !== "") {
                 $formattedAffiliationForInput = Episciences_Paper_AuthorsManager::formatAffiliationForInputRor($affi);
                 $arrayFormOption['affiliations'] = $formattedAffiliationForInput;
+                //avoid future duplicate
+                $acronymAlreadyExisting = Episciences_Paper_AuthorsManager::getAcronymExisting($affi);
+                if ($acronymAlreadyExisting !== '') {
+                    $arrayFormOption['acronymList'] = $acronymAlreadyExisting;
+                }
             }
             $affiForm = Episciences_PapersManager::getAffiliationsForm($arrayFormOption);
 
@@ -618,6 +623,8 @@ class PaperController extends PaperDefaultController
             $jsonAuthorDecoded = json_decode($value['authors'], true, 512, JSON_UNESCAPED_SLASHES | JSON_THROW_ON_ERROR | JSON_UNESCAPED_UNICODE);
         }
         $arrayAffi = [];
+        $acronyms = $request->getPost("affiliationAcronym");
+        $acronyms = explode('||',$acronyms);
         foreach ($affiliations as $key => $affiliation) {
             if ($affiliation !== "") {
                 $affiliation = explode('#', $affiliation);
@@ -625,16 +632,21 @@ class PaperController extends PaperDefaultController
                 $nameRor = ["name" => rtrim($affiliation[0])];
                 $idArray = [];
                 if ((isset($affiliation[1]) && $affiliation[1] !== "") && str_contains(rtrim($affiliation[1]), $rorDomain)) {
+                    $rawstrAcronym = Episciences_Paper_AuthorsManager::setOrUpdateRorAcronym($acronyms, $affiliation[0]);
+                    $strAcronym = Episciences_Paper_AuthorsManager::cleanAcronym($rawstrAcronym);
                     $idArray["id"] = [
                         ['id' => rtrim($affiliation[1]), 'id-type' => "ROR"]
                     ];
+                    if ($strAcronym !== '') {
+                        $idArray["id"][0]['acronym'] = trim($strAcronym);
+                        $nameRor['name'] = Episciences_Paper_AuthorsManager::eraseAcronymInName($nameRor['name'],$rawstrAcronym);
+                    }
                     $arrayAffi[] = array_merge($nameRor, $idArray);
                 } else {
                     $arrayAffi[] = $nameRor;
                 }
 
             }
-
         }
         // avoid space in url to avoid duplicate affiliations
         $currentUrlchecked = '';